Major reorganization
At some places it's not always possible to use curl or wget
(eg: authenticating proxies), and retrieving 45 patches using a browser
is not the most convenient thing to do. So now we provide the list of
patches in the "patches" directory.

Additionally, the build scripts were moved from scripts/ to packages/
and do not have "build-" in their name. The build script provides extra
environment variables such as "MAKE" and "PMAKE" that can be used by these
scripts for normal make and parallel make.

A new "prepare" step is performed before starting the build. It only
downloads the sources if they're not yet here, otherwise reuses the ones
previously downloaded. This makes it easier to populate the sources from
another medium.

The sources are now retrieved using the get_cached_file script from
formilux-techno. Thus the source files can be found in the cache, and
if not available they will populate the cache.