build: fix incorrect function declaration syntax
At many places we have functions declared this way :

   function foo() {
   }

This is only tolerated by bash and is not supported by other shells
like KSH or ASH. The proper declaration is :

   foo() {
   }

Since we may need the compat package to get maximum portability, let's
avoid bashisms. With that change, all packages correctly build with ksh.