build: support having both the fetch path and the file name in the same variable
Most file locations have the same name locally and remotely, and having
to specify the local file name twice is cumbersome. Additionally it
prevents from fetching a list of files. Now we support finding the file
name after the last '/' and optionally forcing it using '|' for URLs
that are constructed in an inadequate form (eg: git).