packages: ensure sudo doesn't use itself
The sudo package requires the system's sudo to set correct permissions.
The issue comes if "::" or ":.:" appears in the PATH variable, as the
file will use itself to copy to the target file. The same issue happens
if the target path is present in the path. So let's first remove the
unexpected paths from the PATH variable, and ensure we avoid as much
as possible to use the wrong utility. By default we now use ${_SUDO}
instead of sudo if set.